Ran into issue. Burger Component reading as Object
there seems to be an issue.  when i use the Burger component outside
the BurgerBuilder component, it works fine. But if used within
BurgerBuilder component and then BurgerBuilder is used to render
the Burger component, the Burger component is read as an object.

Errors says either :
- there is a mismatch in export/import statement.
- mismatch between a named vs default export type.

Need to research further to resolve the issue.