Lecture 249: completed.
- completed till Fixing a Bug..
note that the value property of the selec input type in the order form object in the ContactData.js is empty, i.e.
value: ''... this can create a bug.. as if user does not changes the select.. and hence does not invoke onChangeHandler.. then the value is remain empty stirng and same wil be saved on the database(firebase).

this has been noted by the instructor.. and will be fixed later in the course.

see you in a bit..