Lecture 277 Completed.
Connecting BurgerBuilder Container to our Store.

Note: interesting thing to note is that the BuildControl buttons in the
BurgerControl component pass a ctrl.type to the method ingredientsAdded
and ingredientsRemoved.. which is automatically received by the
dispatching action object, onIngredientAdded and onIngredientRemoved and
thereafter passed to the reducer as argument.

Next: Lecture 278

see you in a bit..